main_repo/requirements.txt
# Bangladesh Data Content Platform — Python Dependencies
# Install with: pip install -r requirements.txt

# ── Agent Framework ─────────────────────────────────────────────────────────
crewai>=0.80.0
crewai-tools>=0.14.0

# ── LLM SDKs ────────────────────────────────────────────────────────────────
anthropic>=0.40.0          # Claude (primary LLM)
litellm>=1.50.0            # Universal LLM router (Claude + Gemini fallback)

# ── Web / HTTP ───────────────────────────────────────────────────────────────
requests>=2.31.0
beautifulsoup4>=4.12.0
httpx>=0.27.0
lxml>=5.0.0                # Fast HTML/XML parser for BeautifulSoup

# ── PDF Processing ───────────────────────────────────────────────────────────
pdfplumber>=0.10.0         # Extract text and tables from PDFs

# ── Web Search ───────────────────────────────────────────────────────────────
duckduckgo-search>=6.0.0   # Free web search, no API key needed

# ── Environment & Config ─────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── Utilities ────────────────────────────────────────────────────────────────
pydantic>=2.0.0            # Data validation (used by CrewAI)
tenacity>=8.2.0            # Retry logic with exponential backoff
rich>=13.0.0               # Pretty terminal output for debugging
